Moved list item controls into its own folder.
