+++ Source/Clients/NostalgicPlayer/AudiusWindow/ListItems/AudiusMusicListItemControl.resx
