+++ Source/Clients/NostalgicPlayer/AudiusWindow/ListItems/AudiusPlaylistListItemControl.resx
